split between minimal and default error handler
