+++ core/base/src/TROOT.cxx
@@ -586,32 +586,6 @@ TROOT *ROOT::Internal::gROOTLocal = ROOT::GetROOT();
 // ROOTDEBUG, or via the debugger.
 Int_t gDebug;
 
-/// Installed as TError handler after gEnv is constructed
-static Int_t DefaultGetErrorIgnoreLevelHandler() {
-   std::string slevel;
-   auto cstrlevel = gEnv->GetValue("Root.ErrorIgnoreLevel", "Print");
-   while (cstrlevel && *cstrlevel) {
-      slevel.push_back(tolower(*cstrlevel));
-      cstrlevel++;
-   }
-
-   if (slevel == "print")
-      return kPrint;
-   else if (slevel == "info")
-      return kInfo;
-   else if (slevel == "warning")
-      return kWarning;
-   else if (slevel == "error")
-      return kError;
-   else if (slevel == "break")
-      return kBreak;
-   else if (slevel == "syserror")
-      return kSysError;
-   else if (slevel == "fatal")
-      return kFatal;
-   return 0;
-}
-
 
 ClassImp(TROOT);
 
@@ -938,7 +912,6 @@ TROOT::~TROOT()
 
       fMessageHandlers->Delete(); SafeDelete(fMessageHandlers);
 
-      ROOT::Internal::SetErrorIgnoreLevelHandler(ROOT::Internal::ErrorIgnoreLevelHandlerFunc_t());
 #ifdef R__COMPLETE_MEM_TERMINATION
       SafeDelete(fCanvases);
       SafeDelete(fTasks);
@@ -968,7 +941,7 @@ TROOT::~TROOT()
 
       // Cleanup system class
       ROOT::Internal::SetErrorSystemMsgHandler(ROOT::Internal::ErrorSystemMsgHandlerFunc_t());
-      ROOT::Internal::SetErrorAbortHandler(ROOT::Internal::ErrorAbortHandlerFunc_t());
+      SetErrorHandler(ROOT::Internal::MinimalErrorHandler);
       delete gSystem;
 
       // ROOT-6022:
@@ -1956,12 +1929,11 @@ void TROOT::InitSystem()
          fprintf(stderr, "Fix this by defining the HOME shell variable\n");
       }
 
-      ROOT::Internal::SetErrorSystemMsgHandler([](){ return gSystem->GetError(); });
-      ROOT::Internal::SetErrorAbortHandler([](){ gSystem->StackTrace(); gSystem->Abort(); });
-
       // read default files
       gEnv = new TEnv(".rootrc");
-      ROOT::Internal::SetErrorIgnoreLevelHandler(DefaultGetErrorIgnoreLevelHandler);
+
+      ROOT::Internal::SetErrorSystemMsgHandler([](){ return gSystem->GetError(); });
+      SetErrorHandler(DefaultErrorHandler);
 
       gDebug = gEnv->GetValue("Root.Debug", 0);
